(#12392) Register puppet eventlog message dll
Now that puppet supports writing to the eventlog, we need to register
the eventlog message dll when puppet installed as MSI (as opposed to
ruby install.rb).

This commit creates a registry key and two values during install that
the Event Viewer uses to localize eventlog messages. During uninstall
the key and its values are removed.

Since the resources in the RegistryEntries components have changed, the
component code must also be changed.

[1] http://msdn.microsoft.com/en-us/library/aa367849(v=vs.85).aspx