Add hiera to Windows packaging.
This commit makes the necessary changes to package hiera with the rest of
puppet and facter. It adds hiera's lib dir to the rubylib and bin dir to the
path in environment.bat and adds a hiera.bat to the bin directory, to allow
hiera to be called from the windows command prompt. It also adds logic to the
rakefile to clone, checkout, and add hiera to the staging area. Finally, it
updates the wording in the installer to remove references to compatibility with
Puppet Enterprise, as no Puppet Enterprise versions exist that are compatible
with Puppet 3.0 yet.