(#16581) Clone ruby from repo instead of zip file
Previously, puppet_for_the_win would download ruby.zip and tools.zip
from dowloads.puppetlabs.com, and bundle them into the MSI. This was
problematic as they were not in source control.

This commit changes the windows rake tasks to clone from the
puppet-win32-ruby repo instead, which contains both ruby & tools. This
eliminated the need for the 7za and rsync utilities for managing the zip
files.

In the process the rake tasks was simplified, particularly the
dependencies among tasks was made explicit.

The UI tasks were removed and can be added back when needed.

The rake tasks now accept a `config` parameter which specifies the
location and ref for each of its downloaded components. This way the build
for FOSS, PE, and CVEs can be kept in source control, instead of in the
jenkins job definition.

Also, the location of puppet's version information was changed (from
puppet.rb to puppet/version.rb), so the code that patched the version
string for Puppet Enterprise would fail as soon as we rebased PE to build
from 2.7.20. This has been fixed so that we check puppet/version.rb first,
but if it's not present, e.g. PE 2.5 based off of Puppet 2.7.12, then we
fall back to the old location.