Maint: Delete downloads that aren't part of the package
Previously, if the downloads directory contained a file/dir that wasn't
a repo we wanted to package in the installer, e.g. ruby.zip left over
from the old way of packaging ruby, then the tasks would copy it to the
stagedir, and then run candle on it, and include it in the package. This
could also happen if you built 3.x puppet, which includes hiera, and then
switched to 2.7.x, causing hiera to be included in the package when it
shouldn't have been.

This commit adds a task to walk each file/dir within the downloads
directory, and remove any that are not repos we want to include in the
package.