(#18046) Enable rubygems support
Previously, puppet used to require rubygems, but this broke environments
that didn't use rubygems. The decision to require rubygems was moved out
of puppet and into the code that loads puppet. This was done for
daemon.bat, but not for the other scripts, e.g. puppet.bat.

This commit enables rubygems support in the environment.bat file which all
of the other bat files source.