Update windows.rake for mcollective packaging
This commit updates windows.rake to handle adding mcollective to the PE 3
packages. Because mcollective will only be there for the enterprise package,
environment.bat can no longer be the same in FOSS and Enterprise, so this
converts environment.bat to an erb template that includes the mcollective and
mcollective_plugins directories when it is an enterprise build.

This also adds a special case for mcollective that moves its plugins
directory into it's own directory in the stagedir, so they can be
independently added to the load path in environment.bat.