Add MCO directories, service, and server.cfg to puppet.wxs
This commit adds the necessary elements to lay down mcollective in PE 3. It
adds the service for the server set to manual start, since mcollective
must be configured by the pe_mcollective module first, and once that
happens the service will be changed to autostart.

It also creates the vardir for logging and the etc directory for the
server.cfg and facts.yaml.