Newer versions of the win32 gems depend on ffi when running under ruby 1.9.
This commit mentions to the user (during installation) that puppet has a
dependency on ffi, which is licensed LGPLv3.

Since the puppet installer is a "Combined Work" containing Artistic 2.0,
ASL 2.0, MIT, and LGPLv3, we "Accompany the Combined Work with a copy of
the GNU GPL and this [LGPLv3] document"[1] to the license file displayed
to the user in the Installer dialog.

[1] http://www.gnu.org/licenses/lgpl.txt