(#22064) lock down facts.d folder
This locks down the facts.d folder on Windows. Without this patch the folder
may be created and have a potential local privilege escalation issue. This
could cause future security concerns. Up until now this has not been
necessary as we have not supported executable external facts on Windows.
This patch should be released in tandem with the executable external facts
support for Windows.

This sets Administrators and SYSTEM to be the only accounts with Write access
to the folder. Users will be given Read/Execute only.  This will set the permissions
on the folder whether it already existed or not. This will apply the
permissions every time the installer is run in the future.