(#22368) allow agent environment
This allows agent environment to be set during the installer and adds
environment=ENV to the puppet.conf file. By default this will set environment
to production.

This brings parity of this feature to Windows as compared with non-Windows OSes (they already have
the ability to specify agent environment during installation).

You can set this by passing it as a MSI property PUPPET_AGENT_ENVIRONMENT.