(#21243) Don't modify expiration policy for existing users
Previously, if an alternate service account was specified, the installer
would change the account to never expire, which is not something we want
to do.

This commit no longer modifies the account. This means if puppet is
configured to run as a domain or local account, and the account expires,
then puppet will not start.