(#22906) remember properties during upgrade
In previous versions of puppet, we did not overwrite existing settings. Now
that we are with the work done in #20281, we have found that the way we are
doing the remember me property doesn't work with the way we were doing
default values. This change fixes the default values to move them to a
customaction that happens just after the commandline / registry search to set
the value. If there is still no value, then the value is set to a default.
This happens prior to the GUI so a value from the registry or supplied on the
command line will show up as the value in the UI boxes.