(#21243) Use computer name as the domain if none specified
Previously, in order to execute the service under a local account, you had
to specify the local computer name for the domain:

    PUPPET_AGENT_ACCOUNT_DOMAIN=%COMPUTERNAME%

and specifying a domain of `.` is not technically valid, though it's what
most users would try to do.

This commit changes the installer so that if the DOMAIN property is not
specified or is specified as '.' on the command line, then the DOMAIN
property will be set to the local computer name.