(#21243) Domain must be NT AUTHORITY for some service accounts
Previously, trying to install the service to run as LocalService or
NetworkService would fail, because those accounts must have a domain
component of NT AUTHORITY. This is unlike the LocalSystem or local users
accounts which are of the form <computername>\LocalSystem.

This commit always sets the PUPPET_AGENT_ACCOUNT_DOMAIN to NT AUTHORITY if
PUPPET_AGENT_ACCOUNT_USER is specified as either LocalService or
NetworkService.

Ideally, I would have liked to create two SetProperty elements to set the
DOMAIN property based on different conditions. But this results in two
custom actions with the same name. The `Action` attribute allows the
custom action name to be specified uniquely, but it is only available in
WiX 3.6, and we're on 3.5. So instead, I create two custom actions, and
invoke them from both install sequences.