(PUP-3003) Exclude util and patch files
This commit switches to using %r to build regexp, and excludes /util/* and
*.patch from files included in the MSI.

It also switches FileUtil.mkdir to mkdir_p, because the former will fail if
the destination directory exists, for example, if you re-run the `build` task
during development. This isn't an issue during automated builds because we
always wipe the workspace before building.