(PUP-3471) Do not expand path variables
Prior to this when resolving the current path to remove the x64bit path in the
x86 installer, we were using a method that expands the path variables. This
means that if someone installs the x86 MSI and they are using variables in
their path environment variable, it would have been resaved as expanded during
installation. This addresses that by querying the registry value directly with
a special switch not to expand environment variables.

We are also opting for a case insensitive regex replace.