(PUP-1802) Use xslt filter to skip ruby.exe in sys.wxs
Prior to this change, the MSI would generate but validation would throw an
error due to ICE30 errors. This was related to a double specification of
ruby.exe. We need to put the file in the component next to where we are
installing the service and quite a bit of control there, but we autogenerate
the sys.wxs file based on heat.exe. With this commit, we are adding a
filters.xslt transform to use when running heat that will automatically skip
just ruby.exe and let everything else through. This will allow us to specify
ruby.exe as part of the ServiceInstall component and not run into ICE30 errors.