(PUP-3843) Remove branding
This removes PE specific branding. It does make the observation that
version is now preferred to the same way that PE formerly did it where
you pass in `PE_VERSION_STRING`. However that has been renamed to
`AGENT_VERSION_STRING` to more accurately reflect what it is used for.

This is still waiting on what happens with `puppet.conf` `graph`
setting in PE.