(MCO-548) Discontinue use of MCO daemon.bat
Currently when mcollective runs as a service on Windows, it executes a
batch script in a cmd.exe process. This adds a layer of unnecessary
complexity when we could call our vendored rubyw directly with the
necessary arguments for the service to run.

This switches over to using rubyw.exe directly and updates filters.xslt
to exclude rubyw.

Paired-with: Ethan J. Brown <ethan@puppetlabs.com>