(maint) Remove erb template
environment.bat was the only file that at one point may have been an erb
template. It no longer is dynamic at all, and we no longer need to
process it as a template. Since it is also the only file that was
treated as a template, we also no longer need to maintain any of the erb
processing automation in this repo