(PUP-4067) Add package manifest and module dirs
In order to support the puppet-agent run in a masterless setting, we
must provide these directories via the puppet agent installation, rather
than the puppet master installation.

This commit also adds a well annotated example environment.conf to the
default production environment that is laid down in the puppet-agent
package so that users will be familiar with what options they have to
customize their environments.