(MCO-597) Ensure mco.bat
MCO should be available as a command from the command line so ensure
that mco.bat is copied to the bin directory to be picked up as a
fragment.

Note that in a later commit we may ship a sample server.cfg for the
sake of end users, as we do with other platforms.  However, we've
delayed that until a later date.