Revert "(maint) Ensure there are no spaces in the msi name"
This reverts commit f6584df2d8f1a9318ae4a8448d9d2a64237743cb.

This commit was originally introduced to deal with issues passing around
variables. For some reason, there was whitespace included in the
environment variable names, which were breaking the wix build. From
testing, it seems like the pipeline has been updated, and we're no
longer seeng the extra spaces. So, this commit doesn't need to be here
anymore.