(MCO-625) Add mco.bat / required environment vars
Drop the use of the mco.bat from the marionette-collective repo
and instead prefer the use of an mco.bat included in PFW. This is
similar to how other bat files are included, e.g. puppet.bat and
facter.bat.

Also inlucde an environment variable for mcollectived.

This reverts changes introduced with MCO-597 in 07ea56ed.