(PA-68) Filter test assemblies for Puppet MSI
Use a WiX xslt filter to ignore the following files:

lth_cat.exe
libfacter_test.exe
libtest.so
libtest1.so
cpp-pcp-client-unittests.exe
pxp-agent-unittests.exe

The first attempt at this removed them from the staging directories
inside ther rake task for building the agent itself. This was decided to
be too heavy handed, as end users may want the test assemblies in
downstream CI pipelines.

The second attempt used WiX filters to ignore the specified files when generating
WiX fragment files from the staging directory. This was a better
approach in that it left the files for anyone who needed them, and
removed them on the packaging step where they were not wanted. However
the filters turned out to be too greedy, and removed hundreds of files
from the MSI.

The third attempt uses a stricter filter that only removes files
matching the names provided. There is too much overlap with words like
'test' to rely on simple matches. The better way to have approached this
is to have an xslt for each product (one for puppet, one for facter,
etc...) however this would require changing the rake tasks significantly
and would probably be superseded by vanagaon.