(PUP-5933) Don't set pluginsync in puppet.conf
As of Puppet 4.4.0, setting the pluginsync option by CLI
or config is deprecated and will log a deprecation warning,
as that behavior will eventually be managed based on whether or not
we are using a cached catalog.

While the setting is deprecated, it still defaults to true,
so removing this setting in puppet.conf will not cause a
change in behavior.