(PUP-5769) Keep Existing Puppet.Conf Settings
In #20281 (167999a13afd622b0bc823650efbe303106b5ad0), we made a
change to allow overriding the current puppet.conf settings. This
came out in Puppet 3.4.0. This allowed being able to pass updated
settings to the installer and having them persist to the conf
file. Unfortunately, that also meant that any changes to the
puppet.conf file outside of install/upgrade time would be ignored
and overwritten by the remembered property values, which is
undesirable.

The conf file should also be looked at as an additional source of
remembered properties to allow keeping those settings when a user
is not overriding those settings during the install/upgrade. To do
this we need a Type 38 Custom Action, that's right embedded
VBScript. It's the only type of custom action that will allow us
to read values from a file and set MSI properties, since we
won't be able to copy the INI file to the Windows directory
to read the values using IniFileSearch (because we won't have
admin permission until we actually start the installation phase).